separate api request functions
